requirements.txt
Django==1.8.5
djangorestframework
requests